Rename zcli:theme command to zcli:themes
