+++ packages/zcli-themes/bin/run.cmd
