+++ packages/zcli-themes/src/lib/getAssets.test.ts
