+++ packages/zcli-themes/src/lib/getAssets.ts
