+++ packages/zcli-themes/src/lib/getManifest.test.ts
