+++ packages/zcli-themes/src/lib/getManifest.ts
