+++ packages/zcli-themes/src/lib/getVariables.ts
