+++ packages/zcli-themes/tests/functional/mocks/base_theme/assets/bike.png
