+++ packages/zcli-themes/tests/functional/mocks/base_theme/manifest.json
