+++ packages/zcli-themes/tests/functional/mocks/base_theme/settings/logo.png
