+++ packages/zcli-themes/tests/functional/mocks/base_theme/zcli.themes.config.json
