+++ packages/zcli/package.json
@@ -17,7 +17,7 @@
     "@oclif/plugin-update": "=3.0.13",
     "@oclif/plugin-warn-if-update-available": "=2.0.20",
     "@zendesk/zcli-apps": "^1.0.0-beta.32",
-    "@zendesk/zcli-theme": "^1.0.0-beta.32",
+    "@zendesk/zcli-themes": "^1.0.0-beta.32",
     "@zendesk/zcli-core": "^1.0.0-beta.32",
     "chalk": "^4.1.2",
     "tslib": "^2.4.0"
@@ -46,7 +46,7 @@
       "@oclif/plugin-not-found",
       "@oclif/plugin-warn-if-update-available",
       "@zendesk/zcli-apps",
-      "@zendesk/zcli-theme"
+      "@zendesk/zcli-themes"
     ],
     "warn-if-update-available": {
       "timeoutInDays": 7,
@@ -56,8 +56,8 @@
       "apps": {
         "description": "manage Zendesk apps workflow"
       },
-      "theme": {
-        "description": "manage Zendesk theme workflow"
+      "themes": {
+        "description": "manage Zendesk themes workflow"
       },
       "profiles": {
         "description": "manage cli user profiles"
